Fix Facebook Login Bug when no system account is present
When there is no system account found, the second login process starts
with Facebook again in the completion process but _authHandler
was being set to nil in the completion process so _authHandler
was not being called after the second login completed. Now
_authHandler is not being set to nil when the second login
starts so it is called at the end of the second login.